[Move]GiftItemInterpreterを移動
