+++ src/main/scala/com/github/unchama/seichiassist/subsystems/seichilevelupgift/bukkit/BukkitGiftItemInterpreter.scala
@@ -4,7 +4,10 @@ import cats.data.Kleisli
 import cats.effect.Sync
 import com.github.unchama.minecraft.actions.OnMinecraftServerThread
 import com.github.unchama.seichiassist.data.{GachaSkullData, ItemData}
-import com.github.unchama.seichiassist.subsystems.seichilevelupgift.domain.Gift
+import com.github.unchama.seichiassist.subsystems.seichilevelupgift.domain.{
+  Gift,
+  GiftItemInterpreter
+}
 import com.github.unchama.seichiassist.subsystems.seichilevelupgift.domain.Gift.Item
 import com.github.unchama.seichiassist.util.Util.grantItemStacksEffect
 import org.bukkit.entity.Player
