+++ src/main/scala/com/github/unchama/seichiassist/subsystems/seichilevelupgift/domain/GiftItemInterpreter.scala
@@ -1,12 +1,8 @@
-package com.github.unchama.seichiassist.subsystems.seichilevelupgift.bukkit
+package com.github.unchama.seichiassist.subsystems.seichilevelupgift.domain
 
 import cats.data.Kleisli
 import cats.effect.Sync
 import com.github.unchama.minecraft.actions.OnMinecraftServerThread
-import com.github.unchama.seichiassist.data.{GachaSkullData, ItemData}
-import com.github.unchama.seichiassist.subsystems.seichilevelupgift.domain.Gift
-import com.github.unchama.seichiassist.subsystems.seichilevelupgift.domain.Gift.Item
-import com.github.unchama.seichiassist.util.Util.grantItemStacksEffect
 import org.bukkit.entity.Player
 
 /**
